pre-LD basic engine work
Right, I'm headed out for a little stroll to get some sunshine before the jam starts.

This is what I have done so far in the engine - notably, absolutely nothing to do with the game that I am going to make for Ludum Dare, because I have no idea what it's going to be right now because the theme hasn't been announced.

This is just some very low-level sprite rendering and object management code. There's also the start of some input handling code as well, but I didn't really get into that much as I have been more focused on getting at least a clean compile on something resembling a game engine that can update and render a bunch of sprites.

I expect I will need to continue working on this engine code throughout the jam. To keep things compartmentalized, I will clone this repo into a new one just for LD55. If there is anything I do there that I'd like to migrate back to this code, I will do that in a pull request or something after the jam.